TIP-860: Move imagine loader from enrich bundle to enrichment bundle
