+++ src/Akeneo/Platform/Bundle/UIBundle/DependencyInjection/PimUIExtension.php
@@ -30,6 +30,7 @@ class PimUIExtension extends Extension
         $loader->load('forms.yml');
         $loader->load('twig.yml');
         $loader->load('providers.yml');
+        $loader->load('services.yml');
 
         $container->setParameter('pim_ui.wrap_class', 'block-wrap');
         $this->placeholdersConfig($config, $container);
