+++ src/Pim/Bundle/EnrichBundle/Resources/config/resolvers.yml
@@ -1,16 +1,7 @@
 parameters:
-    pim_enrich.imagine.loader.flysystem.class: Pim\Bundle\EnrichBundle\Imagine\Loader\FlysystemLoader
     pim_enrich.resolver.locale.class:          Pim\Bundle\EnrichBundle\Resolver\LocaleResolver
 
 services:
-    pim_enrich.imagine.loader.flysystem:
-        class: '%pim_enrich.imagine.loader.flysystem.class%'
-        arguments:
-            - '@akeneo_file_storage.file_storage.filesystem_provider'
-            - ['catalogStorage']
-        tags:
-            - { name: liip_imagine.binary.loader, loader: flysystem_data_loader }
-
     pim_enrich.resolver.locale:
         class: '%pim_enrich.resolver.locale.class%'
         arguments:
